refactor auth and fix drawer menu
