+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "PartyMaker",
+  "name": "partymaker",
   "version": "0.1.0",
   "private": true,
   "description": "PartyMaker app",
@@ -36,6 +36,7 @@
     "jwt-decode": "^2.2.0",
     "lodash": "^4.17.10",
     "lodash-clean": "^2.0.1",
+    "mdi-react": "^4.0.0",
     "moment": "^2.22.2",
     "prop-types": "^15.6.1",
     "react": "^16.4.0",
