+++ src/components/@auth/AuthRoute.jsx
@@ -1,11 +1,10 @@
-/* eslint-disable react/no-did-mount-set-state */
 import React from 'react'
 import PropTypes from 'prop-types'
 import { Redirect, Route } from 'react-router-dom'
 import connector from './connector'
 
 const AuthRoute = ({ auth, ...rest }) => {
-  if (auth.user) {
+  if (auth.token) {
     return <Route {...rest} />
   }
