+++ src/components/Drawer/DrawerMenu/index.jsx
@@ -10,7 +10,7 @@ import Search from '@material-ui/icons/es/Search'
 import Person from '@material-ui/icons/es/Person'
 import Typography from '@material-ui/core/es/Typography/Typography'
 import Grid from '@material-ui/core/es/Grid/Grid'
-import Settings from '@material-ui/icons/es/Settings'
+import Settings from '@material-ui/icons/es/SettingsApplicationsRounded'
 import Background from './Background'
 import connector from './connector'
 import ExitToApp from '@material-ui/icons/es/ExitToApp'
