--- src/components/auth/authComponent/Background.jsx
@@ -1,28 +1,0 @@
-/* eslint-disable react/forbid-prop-types,max-len */
-import React from 'react'
-import PropTypes from 'prop-types'
-
-const url = 'http://localhost:3333/images/summer.jpg'
-const style = {
-  backgroundImage: `url(${url})`,
-  backgroundPosition: 'center',
-  backgroundSize: 'cover',
-  width: '100%',
-  height: '100%',
-  position: 'relative',
-}
-
-const Background = ({ children }) =>
-  <div style={style}>
-    {children}
-  </div>
-
-Background.propTypes = {
-  children: PropTypes.any,
-}
-
-Background.defaultProps = {
-  children: '',
-}
-
-export default Background
